UUID: int2long clean up

+++ app/src/main/java/com/money/manager/ex/servicelayer/CategoryService.java
@@ -68,7 +68,7 @@ public class CategoryService
                 .insert(repo.getUri(), values);
         long id = ContentUris.parseId(result);
 
-        return ((int) id);
+        return id;
     }
 
     public String getCategorySubcategoryName(long categoryId) {
